% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaLong_package.R
\name{.fit_intercept_mv}
\alias{.fit_intercept_mv}
\title{rma.mv engine: builds working V matrix, fits rma.mv, returns tidy inference}
\usage{
.fit_intercept_mv(yi, vi, cluster, rho, small_sample, alpha, method = "REML")
}
\description{
Used when studies contribute multiple effects at the same time point.
}
\keyword{internal}
