% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaLong_package.R
\name{.to_r_scale}
\alias{.to_r_scale}
\title{Convert effect to correlation scale (for ITCV)}
\usage{
.to_r_scale(theta, sy2)
}
\description{
Convert effect to correlation scale (for ITCV)
}
\keyword{internal}
