% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_meta.R
\name{fits}
\alias{fits}
\title{Extract stored fitted model objects}
\usage{
fits(x)
}
\arguments{
\item{x}{An \code{ml_meta} object.}
}
\value{
Named list of fitted model objects, one per estimable time point.
}
\description{
Extract stored fitted model objects
}
