% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_meta.R
\name{tidy}
\alias{tidy}
\title{Tidy an metaLong object into a clean data frame}
\usage{
tidy(x, ...)
}
\arguments{
\item{x}{A \code{ml_sens} or \code{ml_benchmark} object.}

\item{...}{Additional arguments (unused).}
}
\value{
A tidy \code{data.frame}.
}
\description{
Tidy an metaLong object into a clean data frame
}
